requirements.txt
open3d>=0.18.0
numpy>=1.24
scipy>=1.10
pyyaml>=6.0
tqdm>=4.65
pandas>=2.0
tabulate>=0.9
torch>=2.0
# Optional deep-learning backends:
# MinkowskiEngine  -- build from source for FCGF
# geotransformer   -- install from GeoTransformer repo
# OverlapPredator  -- clone prs-eth/OverlapPredator
